requirements.txt
asyncio==3.4.3
asyncpg==0.29.0
nats-py==2.7.2
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
